Base configuration https://g2plot.antv.vision/en/docs/api/plots/bar

Base configuration https://g2plot.antv.vision/en/docs/api/plots/bar

Plot Container

width

height

autoFit

padding

appendPadding

renderer

pixelRatio

limitInPlot

- width optional number default: 400 Set the width of the chart.

- height optional number default: 400 Set the height of the chart.

- autoFit optional boolean default: true Whether the chart automatically adjusts to fit the container. If it is set to true, width and height configuration would fail.

- padding optional number[] | number | 'auto' Set padding value of the canvas. You can also use auto.

- appendPadding optional number[] | number Extra appendPadding value.

- renderer optional string default: canvas Set the render way to canvas or svg.

- pixelRatio optional number default: window.devicePixelRatio Set the pixel ratio of the chart.

- limitInPlot optional boolean Whether clip the Geometry beyond the coordinate system.

Components

https://g2plot.antv.vision/en/docs/api/components/axis

Graphic Style

https://g2plot.antv.vision/en/docs/api/graphic-style

Examples

if (FALSE) {
cfg = list(
  legend = list(position = 'top-left')
)
g2bar(df, x, y, cfg = cfg)

}